+++ src/client/MainWindow.h
@@ -21,6 +21,7 @@ public:
 
 protected:
     void resizeEvent(QResizeEvent *event) override;
+    void showEvent(QShowEvent *event) override;
 
 private:
     void setupScene();